QUEST_BP.tsv [Lines1051-1138]
Ok, I did the line 1051 till line 1138. It's a complete story inside the game lore. A beautiful and moving one indeed. It was a heavy one to translate, so i hope you guys can read it carefully and check if there is any mistakes or typos. There is a lot of 'pending' words too that we are waiting till we decide the best translation. Thank you guys.